pyproject.toml
[tool.poetry]
name = "dingus"
version = "0.6.0"
description = "Command-line Lisk explorer"
authors = ["Alessandro Ricottone <ricott2@gmail.com>"]
license = "MIT"
packages = [
    { include = "dingus" },
]

[tool.poetry.dependencies]
python = "^3.10"
urwid = "^2.1.2"
requests = "^2.26.0"
PyNaCl = "^1.4.0"
protobuf = "^3.19.1"
pyperclip = "^1.8.2"
google = "^3.0.0"
python-socketio = {extras = ["asyncio_client"], version = "^5.5.0"}
pyaes = "^1.6.1"
ECPy = "^1.2.5"
petlib = "^0.0.45"
blspy = "^1.0.15"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[tool.poetry.scripts]
dingus = "dingus.__main__:start"
compile-schemas = "dingus.codec.utils:compile_schemas"

# pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -v --durations=0 --import-mode=importlib"
testpaths = [
    "tests",
    "tests/tree"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
